+++ CHANGES.rst
@@ -13,6 +13,7 @@ Unreleased
   off the warning.
 - :meth:`.user_selectable` to get available subreddit link flairs.
 - Automatic RateLimit handling will support errors with millisecond resolution.
+- :meth:`.pin` to pin the submission on user's profile
 
 **Fixed**